requirements.txt
alembic==1.7.7
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==1.1.4
Flask-Cors==3.0.10
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.4.0
ipykernel==6.15.0
ipython==8.4.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==2.11.3
jupyter-client==7.3.4
jupyter-core==4.10.0
Mako==1.2.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.8.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==23.2.0
requests==2.28.0
six==1.16.0
SQLAlchemy==1.4.32
stack-data==0.3.0
tornado==6.1
traitlets==5.3.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==1.0.1
zipp==3.7.0
